Delete folder fix

+++ lib/commands/thor/module.rb
@@ -68,9 +68,11 @@ module DTK::Client
     method_option :force, :aliases => '-y', :type => :boolean, :default => false
     method_option :purge, :aliases => '-p', :type => :boolean, :default => false
     def delete(context_params)
+      module_id, module_info, module_location, modules_path = nil, nil, nil, nil
       component_module_id, force_delete = context_params.retrieve_arguments([:option_1!, :option_2],method_argument_names)
       # add component_module_id/name required by info method
       context_params.add_context_to_params("module", "module", component_module_id)
+      module_info = info(context_params)
       
       unless (options.force? || force_delete)
         # Ask user if really want to delete component module and all items contained in it, if not then return to dtk-shell without deleting
@@ -84,21 +86,23 @@ module DTK::Client
       return response unless response.ok?
       module_name = response.data(:module_name)
       Helper(:git_repo).unlink_local_clone?(:component_module,module_name)
+      
+      # when changing context send request for getting latest modules instead of getting from cache
+      @@invalidate_map << :module_component
 
       # delete local module directory
       if options.purge?
-        module_id         = nil
-        module_location   = nil
-        module_info       = info(context_params)
-        module_id         = module_info["data"]["display_name"] if module_info["status"] == "ok"
-        modules_path      = OsUtil.module_clone_location(::Config::Configuration.get(:module_location))
-        module_location   = "#{modules_path}/#{module_id}" unless (module_id.nil? || module_id.empty?)
+        raise DTK::Client::DtkValidationError, "Unable to delete local directory. Message: #{module_info['errors'].first['message']}." unless module_info["status"] == "ok"
+        module_id       = module_info["data"]["display_name"]
+        modules_path    = OsUtil.module_clone_location(::Config::Configuration.get(:module_location))
+        module_location = "#{modules_path}/#{module_id}" unless (module_id.nil? || module_id.empty?)
 
-        FileUtils.rm_rf("#{module_location}") if File.directory?(module_location)
+        unless (module_location.nil? || ("#{modules_path}/" == module_location))
+          FileUtils.rm_rf("#{module_location}") if File.directory?(module_location)
+        end
       end
 
-      # when changing context send request for getting latest modules instead of getting from cache
-      @@invalidate_map << :module_component
+      return response
     end
 
     #
